requirements.txt
discord.py>=2.5.0
python-dotenv>=1.0.0
aiohttp>=3.9.1
asyncpg>=0.29.0
psutil>=5.9.0
